# More efficient implementation of `FastStringReader`